+++ R/convert4crosstab.R
@@ -12,12 +12,12 @@ function(cross, chr)
 {
     if(!missing(chr)) cross <- cross[chr,]
 
-    crosstype <- class(cross)[1] # f2, bc, etc.
+    crosstype <- crosstype(cross) # f2, bc, etc.
     mnames <- qtl::markernames(cross)
 
     # chr type for each marker ("A" or "X")
-    chrtype <- rep(sapply(cross$geno, class), qtl::nmar(cross))
-    names(chrtype) <- mnames
+    chr_type <- rep(sapply(cross$geno, chrtype), qtl::nmar(cross))
+    names(chr_type) <- mnames
 
     # for conversion with X chr
     cross.attr <- attributes(cross)
@@ -25,15 +25,15 @@ function(cross, chr)
 
     # names of the genotype categories; making missing values the last category
     genocat <- vector("list", 0)
-    if(any(chrtype == "A"))
+    if(any(chr_type == "A"))
         genocat$A <- c(qtl::getgenonames(crosstype, "A", cross.attr=cross.attr), "N/A")
-    if(any(chrtype == "X"))
+    if(any(chr_type == "X"))
         genocat$X <- c(qtl::getgenonames(crosstype, "X", expandX = "standard", sexpgm=sexpgm,
                                          cross.attr=attributes(cross)), "N/A")
 
     geno <- qtl::pull.geno(cross)
-    if(any(chrtype=="A")) {
-        genoA <- geno[,chrtype=="A"]
+    if(any(chr_type=="A")) {
+        genoA <- geno[,chr_type=="A"]
         # replace missing values with number for last category
         genoA[is.na(genoA)] <- length(genocat$A)
         # check for unexpected codes
@@ -43,10 +43,10 @@ function(cross, chr)
             warning("Unexpected genotype codes for markers ", paste(badmar, collapse=" "))
             genoA[genoA < 1 | genoA > length(genocat$A)] <- length(genocat$A)
         }
-        geno[,chrtype=="A"] <- genoA
+        geno[,chr_type=="A"] <- genoA
     }
-    if(any(chrtype=="X")) {
-        genoX <- geno[,chrtype=="X"]
+    if(any(chr_type=="X")) {
+        genoX <- geno[,chr_type=="X"]
         # convert from 1/2 to numbers for like AA/AB/BB/AY/BY
         genoX <- qtl::reviseXdata(crosstype, expandX="standard", sexpgm=sexpgm, geno=genoX,
                                   cross.attr=cross.attr)
@@ -59,10 +59,10 @@ function(cross, chr)
             warning("Unexpected genotype codes for markers ", paste(badmar, collapse=" "))
             genoX[genoX < 1 | genoX > length(genocat$X)] <- length(genocat$X)
         }
-        geno[,chrtype=="X"] <- genoX
+        geno[,chr_type=="X"] <- genoX
     }
     # make it a list
     geno <- as.list(as.data.frame(geno))
 
-    list(geno=geno, genocat=genocat, chrtype=as.list(chrtype))
+    list(geno=geno, genocat=genocat, chrtype=as.list(chr_type))
 }